phase5_crewai/requirements.txt
# ============================================
# Phase 5: CrewAI - 依赖包
# ============================================
# 安装: pip install -r requirements.txt

# ============================================
# CrewAI 核心
# ============================================
crewai==0.1.26
crewai-tools==0.0.15

# ============================================
# OpenAI 集成
# ============================================
langchain-openai==0.0.5
openai==1.10.0

# ============================================
# 工具依赖
# ============================================
duckduckgo-search==5.3.0
google-search-results==2.4.2
requests==2.31.0

# ============================================
# 其他依赖
# ============================================
python-dotenv==1.0.0
jupyter==1.0.0
ipykernel==6.29.2

# ============================================
# 开发工具
# ============================================
black==23.12.1
isort==5.13.2
flake8==7.0.0
